pyproject.toml
[project]
name = "qa-backend"
version = "0.1.0"
description = "Project for testing backend of my copy of the pokemon api"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
]


[dependency-groups]
dev = [
    "ruff==0.15.8",
    "pre-commit==4.5.1",
    "pytest>=9.0.2",
    "pyhamcrest>=2.1.0",
    "requests>=2.33.1",
    "psycopg[binary]>=3.3.3",
]

[tool.ruff]
line-length = 88          
target-version = "py312"
src = ["app"]    

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear (catches common bugs)
    "C4",   # flake8-comprehensions (cleaner list/dict comprehensions)
    "UP",   # pyupgrade (modernize syntax)
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long — handled by formatter
    "B008",   # Do not perform function calls in default args (common in FastAPI)
]
unfixable = ["B"]   # Bugbear fixes can be destructive — review manually

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
